+++ source/index/flat.hpp
@@ -2,7 +2,7 @@
 #define SOILLIB_INDEX_FLAT
 
 #include <soillib/soillib.hpp>
-#include <soillib/core/yield.hpp>
+#include <soillib/util/yield.hpp>
 #include <soillib/core/types.hpp>
 
 #include <vector>
